backend/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy
SQLAlchemy
psycopg2-binary
alembic
requests
beautifulsoup4
gunicorn
python-dotenv